backend/requirement.txt
asgiref==3.11.0
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
Django==5.2.10
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
lxml==6.0.2
pdfminer.six==20260107
pillow==12.1.0
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
PyPDF2==3.0.1
python-docx==1.2.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
requests
python-dotenv
weasyprint